Fix parameter for mysql_install_db
According to https://dev.mysql.com/doc/refman/5.7/en/mysql-install-db.html,
the parameter `--ldata` was removed in mysql 5.7.5. Since it was just a
synonym for `--datadir`, we use that parameter instead.

This fixes the creation of a mysql database for the
`development/ubuntu/xenial` case, where the mysql version is already too
recent to have `--ldata` (currently Ver 5.7.23).